Created various modifier trackers.
Things are set up right now so that concrete implementations of classes cannot be exposed publicly.  Only the GetTotal() methods of these classes can be exposed publicly (I expect all mutation of these classes to occur from interaction with classes within the .Core.Domain library.)  I may need to change this behavior later once other classes are ironed out and the public interface for other objects which use ModifierTrackers is better understood.

This check-in lacks unit tests.  Very few tests should be necessary, since most of these classes contain no logic of their own.

Notably, there is no tracker for size bonuses.  This is because one's size bonus is the result of a calculation based on size, so there's no data to store and aggregate.  If I find an exception to this, I'll add a SizeBonusTracker:NonStackingModifierTracker class later.