Removed Public access to the constructor so that Public callers cannot easily get a concrete instance of Character.  (The factory function returns a reference to ICharacter, which is all a Public caller should need anyway.)  The actual constructor is Internal so that unit tests can get access to Internal access methods and properties of a Character without casting (remember that the unit testing project has Internal access.)
Added an immutable Level value to the Character.  This is validated, since this code is client-facing.  Thinking ahead a little bit, the Level field was made immutable because allowing levels to change may introduce problems later on, such as this situation:
1) Character level is set to 9
2) A skill's ranks is set to 12 (valid)
3) The character's level is set to 1.  (Now the character's ranks aren't valid, and we have a problem!)
Making the value immutable means that other parts of the character (like skill ranks) don't have to listen for an OnCharacterLevelChanged event and revalidate any time the level is changed.  This keeps the code leaner (less to write) and more performant (not spending so much CPU and memory on validation).