Safety check-in.
I created a concept of a "Registered Spell"-- that is, a spell which is associated with a character.  This allows the character's stats to be applied to a particular spell so things like caster level and DC can be calculated.  When a new spell is registered to a character, an event is raised so that feats like Spell Focus can see that a spell has been registered and apply themselves accordingly.  It is my intention than only registered spells can be prepared or added to spells known.

I added the Moq testing framework to make testing simpler.  Depending on how things go, I may need to revisit which classes are sealed (I seal classes by default to save time; I firmly believe that classes should be designed with inheritance in mind or that inheritance should be prevented) and how the interfaces look.

I still need to test the SpellRegister class.