Renamed a class for consistency.
Created a Spellbook class and ISpellbook interface.  The Spellbook class is designed with inheritance in mind (such as for a Blessed Book, which hasn't been implemented yet).  It is my intention that a spellbook will be referenced two places on the Character: once for the purposes of noting what abilities the character has access to (Character.Spellbook) and once for the purposes of noting treasure (Character.GetInventory() or something of the like).