Simplified Spell constructors by replacing constructor arguments with abstract properties.
Created the first feat, SpellFocus, which increases the DC of spells which share the same school by one.  This necessitated a property on (I)RegisteredSpell so that this kind of bonus could be tracked.  Because ModifierTrackers are not exposed publicly, this meant I had to write a public method which wrapped to call to the ModifierTracker which stores the data.

The "unit test" which shows that a Feat can listen to the OnSpellRegistered event and modify the newly registered spell strikes me as a bit too complex for a unit test.  Perhaps I'll create a new project for complex tests like these.