Refactored modifier trackers such that they are able to store dynamic modifiers:
The only reason that IModifierTracker was never intended to be exposed publicly before was because other objects may need to override specific behaviors, such as the Swim skill having a +8 racial bonus if the character has a Swim speed.  Now that the Swim skill's racial IModifierTracker can store this sort of calculation directly, it is not necessary for the Swim skill to have to track this calculation itself.

Folded ITotalable into IModifierTracker and made IModifierTracker public.  This is so that IModifierTracker objects can be exposed directly as members on more complex objects and reduce how much testing is necessary.